Add convertion logic
