+++ src/device.jl
@@ -124,37 +124,26 @@ Return a `LuxCPUDevice` object which can be used to transfer data to CPU.
 @inline get_cpu_device() = LuxCPUDevice()
 
 (::LuxCPUDevice)(x) = fmap(x -> adapt(LuxCPUAdaptor(), x), x; exclude=_isleaf)
+(::LuxCUDADevice)(x) = fmap(x -> adapt(LuxCUDAAdaptor(), x), x; exclude=_isleaf)
+(::LuxAMDGPUDevice)(x) = fmap(x -> adapt(LuxAMDGPUAdaptor(), x), x; exclude=_isleaf)
 
-"""
-    cpu(x)
+# Adapt Interface
+abstract type LuxDeviceAdaptor end
 
-Transfer `x` to CPU.
+struct LuxCPUAdaptor <: LuxDeviceAdaptor end
+struct LuxCUDAAdaptor <: LuxDeviceAdaptor end
+struct LuxAMDGPUAdaptor <: LuxDeviceAdaptor end
 
-!!! warning
-
-    This function has been deprecated. Use `get_cpu_device` instead.
-"""
-function cpu(x)
-    Base.depwarn("This function has been deprecated and will be removed in v0.6. Use `get_cpu_device` instead.",
-        :cpu)
-    device = get_cpu_device()
-    return device(x)
+function adapt_storage(::LuxCPUAdaptor,
+    x::Union{AbstractRange, SparseArrays.AbstractSparseArray})
+    return x
 end
+adapt_storage(::LuxCPUAdaptor, x::AbstractArray) = adapt(Array, x)
+adapt_storage(::LuxCPUAdaptor, rng::AbstractRNG) = rng
 
-"""
-    gpu(x)
+_isbitsarray(::AbstractArray{<:Number}) = true
+_isbitsarray(::AbstractArray{T}) where {T} = isbitstype(T)
+_isbitsarray(x) = false
 
-Transfer `x` to GPU determined by the backend set using `Lux.gpu_backend!`.
-
-!!! warning
-
-    This function has been deprecated. Use `get_gpu_device` instead. Using this function inside
-    performance critical code will cause massive slowdowns due to type inference failure.
-"""
-function gpu(x)
-    @warn "Using this function inside performance critical code will cause massive slowdowns due to type inference failure. Please update your code to use `get_gpu_device` API." maxlog=1
-    Base.depwarn("This function has been deprecated and will be removed in v0.6. Use `get_gpu_device` instead.",
-        :gpu)
-
-    return (get_gpu_device())(x)
-end
+_isleaf(::AbstractRNG) = true
+_isleaf(x) = _isbitsarray(x) || Functors.isleaf(x)
